# src/main/java/de/felixhertweck/seatreservation/supervisor/dto/CheckInProcessRequestDTO.java
/*
 * #%L
 * seat-reservation
 * %%
 * Copyright (C) 2025 Felix Hertweck
 * %%
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * #L%
 */
package de.felixhertweck.seatreservation.supervisor.dto;

import java.util.List;

/**
 * DTO for check-in process request with separate lists for check-ins and cancellations.
 *
 * @param checkIn list of reservation IDs for check-in
 * @param cancel list of reservation IDs for cancellation
 */
public record CheckInProcessRequestDTO(List<Long> checkIn, List<Long> cancel) {}