# webapp/docs/openapi.json
{
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/supervisor/checkin/info" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CheckInInfoRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK - Check-in information retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CheckInInfoResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "400" : {
            "description" : "Bad Request - Invalid input parameters"
          },
          "404" : {
            "description" : "Token not found"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Get Check In Info",
        "tags" : [ "Check In Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "SUPERVISOR", "ADMIN", "MANAGER" ]
        } ]
      }
    },
    "/api/supervisor/checkin/info/{username}" : {
      "post" : {
        "parameters" : [ {
          "name" : "username",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK - Check-in information processed successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CheckInInfoResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "404" : {
            "description" : "User or Reservation not found"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Process Check In Info By Username",
        "tags" : [ "Check In Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "SUPERVISOR", "ADMIN", "MANAGER" ]
        } ]
      }
    },
    "/api/supervisor/checkin/process" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CheckInProcessRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content - Successfully processed"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "400" : {
            "description" : "Bad Request - Invalid input parameters"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Process Check In",
        "tags" : [ "Check In Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "SUPERVISOR", "ADMIN", "MANAGER" ]
        } ]
      }
    },
    "/api/supervisor/checkin/usernames" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK - Usernames retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "Get Usernames With Reservations",
        "tags" : [ "Check In Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "SUPERVISOR", "ADMIN", "MANAGER" ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CheckInInfoRequestDTO" : {
        "type" : "object",
        "required" : [ "userId", "eventId" ],
        "properties" : {
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "eventId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "checkInTokens" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "CheckInInfoResponseDTO" : {
        "type" : "object",
        "properties" : {
          "reservations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LiveReservationResponseDTO"
            }
          },
          "user" : {
            "$ref" : "#/components/schemas/UserDTO"
          }
        }
      },
      "CheckInProcessRequestDTO" : {
        "type" : "object",
        "properties" : {
          "checkIn" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "cancel" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }
      },
...
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
...
      "LiveReservationResponseDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "eventId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "seat" : {
            "$ref" : "#/components/schemas/SeatDTO"
          },
          "reservationDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "status" : {
            "$ref" : "#/components/schemas/ReservationStatus"
          },
          "liveStatus" : {
            "$ref" : "#/components/schemas/ReservationLiveStatus"
          },
          "reservationDate" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
...
      "ReservationLiveStatus" : {
        "type" : "string",
        "enum" : [ "CHECKED_IN", "CANCELLED", "NO_SHOW" ]
      },
...
      "ReservationStatus" : {
        "type" : "string",
        "enum" : [ "RESERVED", "BLOCKED" ]
      },
      "SeatDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "seatNumber" : {
            "type" : "string"
          },
          "seatRow" : {
            "type" : "string"
          },
          "locationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "xCoordinate" : {
            "type" : "integer",
            "format" : "int32"
          },
          "yCoordinate" : {
            "type" : "integer",
            "format" : "int32"
          },
          "entrance" : {
            "type" : "string"
          }
        }
      },
...
      "UserDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "username" : {
            "type" : "string"
          },
          "firstname" : {
            "type" : "string"
          },
          "lastname" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "emailVerified" : {
            "type" : "boolean"
          },
          "emailVerificationSent" : {
            "type" : "boolean"
          },
          "roles" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "type" : "string"
            }
          },
          "tags" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT",
        "description" : "Authentication"
      }
    }
  },
...
  "info" : {
    "title" : "Seat Reservation API",
    "version" : "0.0.7",
    "description" : "The official REST API for Seat Reservation.",
    "contact" : {
      "name" : "FelixHertweck GitHub ",
      "url" : "https://github.com/FelixHertweck"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}